start adding title component
