--- src/views/TicketModal/styles.module.scss
@@ -1,3 +1,0 @@
-.ticket-modal-header {
-    display: flex;
-}
